tweaks based on reddit feedback

+++ docker-deployment/main.py
@@ -9,45 +9,44 @@ logging.basicConfig(level=logging.INFO)
 
 
 def handler(event, context):
+    running_outside_of_lambda = os.environ.get("AWS_LAMBDA_FUNCTION_NAME") is None
     print(event)
+    # Send ol' Benjamin an email to let him know that people are using the service
     try:
         boto3.client("sns").publish(
             TopicArn=os.environ["SNS_TOPIC_ARN"],
-            Message=f"Running tabroom_summary for {event['tournament']}",
+            Message=f"Running tabroom_summary for {event['tournament']}; requested school is {event['school']}",
         )
     except Exception:
         logging.error("Error publishing to SNS")
     # This function will query Tabroom.com for the tournament results and then
     # return a summary of the results using ChatGPT.
     # This Lambda will run asynchronously and will not return the results directly.
-    debug = event.get("debug", False)
-    try:
-        open_ai_key_secret_name = os.environ["OPEN_AI_KEY_SECRET_NAME"]
-    except KeyError:
-        open_ai_key_secret_name = "openai_auth_key"
     tournament_id = event["tournament"]
-    if debug:
+    if running_outside_of_lambda:
         # For debugging, we will run the function locally and return the results to the user
         response = tabroom_summary.main(
             all_schools=True,
             tournament_id=tournament_id,
             open_ai_key_path=event["open_ai_key_path"],
-            debug=debug,
             read_only=event.get("read_only", False),
         )
-    response = tabroom_summary.main(
-        all_schools=True,
-        tournament_id=tournament_id,
-        open_ai_key_secret_name=open_ai_key_secret_name,
-        debug=debug,
-        read_only=event.get("read_only", False),
-    )
+    else:
+        open_ai_key_secret_name = os.environ.get(
+            "OPEN_AI_KEY_SECRET_NAME", "openai_auth_key"
+        )
+        response = tabroom_summary.main(
+            all_schools=True,
+            tournament_id=tournament_id,
+            open_ai_key_secret_name=open_ai_key_secret_name,
+            read_only=event.get("read_only", False),
+        )
     # If we're not in Lambda, assume we're in Windows
-    if os.environ.get("AWS_LAMBDA_FUNCTION_NAME") is None:
+    if running_outside_of_lambda:
         # Make the directories as needed
         for school_name in response.keys():
             os.makedirs(f"{tournament_id}/{school_name}", exist_ok=True)
-            if not event.get("read_only", False):
+            if event.get("read_only", False) is True:
                 if "full_response" in response[school_name]:
                     with open(f"{tournament_id}/{school_name}/results.txt", "w") as f:
                         f.write(response[school_name]["full_response"])
@@ -59,14 +58,15 @@ def handler(event, context):
         s3_client = boto3.client("s3")
         bucket_name = os.environ["DATA_BUCKET_NAME"]
         for school_name in response.keys():
-            if "full_response" not in response[school_name]:
-                logging.warning(f"No GPT response found for {school_name}")
-                continue
-            s3_client.put_object(
-                Body=response[school_name]["full_response"],
-                Bucket=bucket_name,
-                Key=f"{tournament_id}/{school_name}/results.txt",
-            )
+            if event.get("read_only", False) is True:
+                if "full_response" not in response[school_name]:
+                    logging.warning(f"No GPT response found for {school_name}")
+                    continue
+                s3_client.put_object(
+                    Body=response[school_name]["full_response"],
+                    Bucket=bucket_name,
+                    Key=f"{tournament_id}/{school_name}/results.txt",
+                )
             if "gpt_prompt" in response[school_name]:
                 s3_client.put_object(
                     Body=response[school_name]["gpt_prompt"],
@@ -112,7 +112,6 @@ if __name__ == "__main__":
     open_ai_key_path = args.open_ai_key_path
     event = {
         "tournament": tournament_id,  # "30799",  # "29810",  # "20134",
-        "debug": True,  # Will always be true for local testing
         "open_ai_key_path": open_ai_key_path,
         "read_only": read_only,
     }
